requirements.txt
OpenAI==1.25.1
boto3==1.24.89
botocore==1.27.89
click==8.1.3
Flask==2.2.2
flask_cors==4.0.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pika==1.3.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
requests
Werkzeug==2.2.2
requests==2.28.1
pymongo==4.7.2
psutil==5.9.0

Flask-Log-Request-ID==0.10.0
pillow
pillow-heif
opencv-python
pikepdf
pypdf
gunicorn
nltk